+++ Packages/BrandonUtils/Standalone/Clerical/Saving/SaveData.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace BrandonUtils.Standalone.Clerical.Saving {
+    public abstract class SaveData : ISaveData {
+        public DateTime LastSaveTime { get; }
+        public DateTime LastLoadTime { get; }
+    }
+}